pyproject.toml
[project]
name = "pydantic-ai-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "devtools>=0.12.2",
    "ipykernel>=6.29.5",
    "logfire>=2.7.1",
    "nest-asyncio>=1.6.0",
    "pydantic-ai[logfire]>=0.0.13",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.8.3",
]
